+++ core/rend/vulkan/vk_context_lr.h
@@ -44,7 +44,7 @@ public:
 	vk::Device GetDevice() const { return device; }
 	vk::PipelineCache GetPipelineCache() const { return *pipelineCache; }
 	vk::DescriptorPool GetDescriptorPool() const { return *descriptorPool; }
-	size_t GetSwapChainSize() const { u32 m = retro_render_if->get_sync_index_mask(retro_render_if->handle); u32 n = 1; while (m >>= 1) n++; return n; }
+	u32 GetSwapChainSize() const { u32 m = retro_render_if->get_sync_index_mask(retro_render_if->handle); u32 n = 1; while (m >>= 1) n++; return n; }
 	int GetCurrentImageIndex() const { return retro_render_if->get_sync_index(retro_render_if->handle); }
 	// FIXME that's not quite correct
 	void WaitIdle() const { retro_render_if->wait_sync_index(retro_render_if->handle); }
